# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Anime & Film streaming site
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type AnimeCardContentType =
  (typeof AnimeCardContentType)[keyof typeof AnimeCardContentType];

export const AnimeCardContentType = {
  anime: "anime",
  series: "series",
  film: "film",
} as const;

export interface AnimeCard {
  title: string;
  slug: string;
  poster?: string;
  type?: string;
  status?: string;
  rating?: string;
  episode?: string;
  contentType?: AnimeCardContentType;
}

export type AnimeListResponsePagination = {
  currentPage: number;
  hasNext: boolean;
  hasPrev: boolean;
};

export interface AnimeListResponse {
  data: AnimeCard[];
  pagination: AnimeListResponsePagination;
}

export interface HomeData {
  featured: AnimeCard[];
  latest: AnimeCard[];
  popular: AnimeCard[];
  ongoing: AnimeCard[];
}

export interface Genre {
  name: string;
  slug: string;
  count?: string;
}

export interface GenreListResponse {
  data: Genre[];
}

export interface ScheduleDay {
  day: string;
  anime: AnimeCard[];
}

export interface ScheduleResponse {
  schedule: ScheduleDay[];
}

export type AnimeDetailEpisodeListItem = {
  title: string;
  slug: string;
  date?: string;
};

export interface AnimeDetail {
  title: string;
  alternativeTitle?: string;
  poster: string;
  synopsis: string;
  genres?: string[];
  status?: string;
  type?: string;
  rating?: string;
  episodes?: string;
  duration?: string;
  studio?: string;
  released?: string;
  episodeList?: AnimeDetailEpisodeListItem[];
}

export type EpisodeDataServersItem = {
  name: string;
  post?: string;
  nume?: string;
  type?: string;
};

export interface EpisodeData {
  title: string;
  servers: EpisodeDataServersItem[];
  embedUrl?: string;
  prevEpisode?: string;
  nextEpisode?: string;
  animeSlug?: string;
  animeTitle?: string;
}

export interface ServerResponse {
  embedUrl: string;
}

export type SearchAnimeParams = {
  /**
   * Search query
   */
  q: string;
  page?: number;
};

export type GetAnimeListParams = {
  /**
   * Category type
   */
  category: GetAnimeListCategory;
  page?: number;
};

export type GetAnimeListCategory =
  (typeof GetAnimeListCategory)[keyof typeof GetAnimeListCategory];

export const GetAnimeListCategory = {
  animedonghua: "animedonghua",
  film: "film",
  series: "series",
  tvshow: "tvshow",
  others: "others",
  ongoing: "ongoing",
  completed: "completed",
  populer: "populer",
  latest: "latest",
  update: "update",
  "all-anime": "all-anime",
  "all-anime-reverse": "all-anime-reverse",
} as const;

export type GetCatalogParams = {
  title?: string;
  page?: number;
  order?: GetCatalogOrder;
  type?: GetCatalogType;
  status?: GetCatalogStatus;
};

export type GetCatalogOrder =
  (typeof GetCatalogOrder)[keyof typeof GetCatalogOrder];

export const GetCatalogOrder = {
  update: "update",
  title: "title",
  rating: "rating",
} as const;

export type GetCatalogType =
  (typeof GetCatalogType)[keyof typeof GetCatalogType];

export const GetCatalogType = {
  TV: "TV",
  Movie: "Movie",
  OVA: "OVA",
  Special: "Special",
  ONA: "ONA",
  Music: "Music",
} as const;

export type GetCatalogStatus =
  (typeof GetCatalogStatus)[keyof typeof GetCatalogStatus];

export const GetCatalogStatus = {
  Currently_Airing: "Currently Airing",
  Finished_Airing: "Finished Airing",
  Not_yet_aired: "Not yet aired",
} as const;

export type GetByGenreParams = {
  page?: number;
};

export type GetScheduleParams = {
  day?: GetScheduleDay;
};

export type GetScheduleDay =
  (typeof GetScheduleDay)[keyof typeof GetScheduleDay];

export const GetScheduleDay = {
  senin: "senin",
  selasa: "selasa",
  rabu: "rabu",
  kamis: "kamis",
  jumat: "jumat",
  sabtu: "sabtu",
  minggu: "minggu",
} as const;

export type GetAnimeDetailParams = {
  /**
   * Content type for proper URL construction
   */
  type: GetAnimeDetailType;
};

export type GetAnimeDetailType =
  (typeof GetAnimeDetailType)[keyof typeof GetAnimeDetailType];

export const GetAnimeDetailType = {
  anime: "anime",
  series: "series",
  film: "film",
} as const;

export type GetServerParams = {
  post: string;
  nume: string;
  type: string;
};