# lib/api-zod/src/generated/types/animeCard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Anime & Film streaming site
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeCardContentType } from "./animeCardContentType";

export interface AnimeCard {
  title: string;
  slug: string;
  poster?: string;
  type?: string;
  status?: string;
  rating?: string;
  episode?: string;
  contentType?: AnimeCardContentType;
}