# lib/api-zod/src/generated/types/animeDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Anime & Film streaming site
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeDetailEpisodeListItem } from "./animeDetailEpisodeListItem";

export interface AnimeDetail {
  title: string;
  alternativeTitle?: string;
  poster: string;
  synopsis: string;
  genres?: string[];
  status?: string;
  type?: string;
  rating?: string;
  episodes?: string;
  duration?: string;
  studio?: string;
  released?: string;
  episodeList?: AnimeDetailEpisodeListItem[];
}