# lib/api-zod/src/generated/types/animeListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Anime & Film streaming site
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeCard } from "./animeCard";
import type { AnimeListResponsePagination } from "./animeListResponsePagination";

export interface AnimeListResponse {
  data: AnimeCard[];
  pagination: AnimeListResponsePagination;
}