# lib/api-zod/src/generated/types/scheduleResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Anime & Film streaming site
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleDay } from "./scheduleDay";

export interface ScheduleResponse {
  schedule: ScheduleDay[];
}